+++ lib/kettle/family.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require "version_gem"
-
 require_relative "family/version"
 require_relative "family/member"
 require_relative "family/command_result"
@@ -10,8 +8,6 @@ require_relative "family/command_runner"
 require_relative "family/readiness_check"
 require_relative "family/changelog_check"
 require_relative "family/git_status"
-require_relative "family/version_bump"
-require_relative "family/dependency_floor"
 require_relative "family/branch_lane_audit"
 require_relative "family/config"
 require_relative "family/branch_target_config"
@@ -28,8 +24,4 @@ module Kettle
   module Family
     class Error < StandardError; end
   end
-end
-
-Kettle::Family::Version.class_eval do
-  extend VersionGem::Basic
 end